requirements.txt
openai
pandas
sqlalchemy
fastapi
uvicorn
numpy
scikit-learn
python-dotenv